+++ src/Produtos/Coluna.java
@@ -3,7 +3,7 @@ package Produtos;
 public class Coluna extends Viga implements MemoriaCalculo{
 
 
-    public Coluna(int quantidadeItens, int quantidadeBarradeFerro, BitolaBarraFerroEnum mm, float metragem, double medidaEstribo1, double medidaEstribo2, double aCadaQuanto, float dobraPonta) {
+    public Coluna(int quantidadeItens, int quantidadeBarradeFerro, BitolaBarraFerroEnum mm, double metragem, double medidaEstribo1, double medidaEstribo2, double aCadaQuanto, float dobraPonta) {
         super(quantidadeItens, quantidadeBarradeFerro, mm, metragem, medidaEstribo1, medidaEstribo2, aCadaQuanto, dobraPonta);
     }
